raed push

+++ _Tunisia_Charity_Land_Esprit/src/Service/OrderService.java
@@ -62,34 +62,36 @@ public class OrderService {
         // execute the preparedstatement
         preparedStmt.execute();
     }
-      public void suppOrder(int x) throws SQLException
+      public void suppOrder(int x, int y) throws SQLException
     {
-        String requete="DELETE FROM `orders` WHERE `product_id`=?;";
+        String requete="DELETE FROM `orders` WHERE `product_id`=? and user_id=?;";
           PreparedStatement preparedStatement;
         try {
             preparedStatement = connection.prepareStatement(requete);
             preparedStatement.setInt(1, x);
+            preparedStatement.setInt(2, y);
             preparedStatement.executeUpdate();
         } catch (SQLException ex) {
             System.out.println(ex);
         }
     }
-    public void modOrder (int x, int a) throws SQLException
+    public void modOrder (int x, int a,int y) throws SQLException
     {
-        String requete="UPDATE `orders` SET `quantity`=? WHERE `product_id`=? ;";
+        String requete="UPDATE `orders` SET `quantity`=? WHERE `product_id`=? and user_id=? ;";
           PreparedStatement preparedStatement;
         try {
             preparedStatement = connection.prepareStatement(requete);
             preparedStatement.setInt(1, x);
             preparedStatement.setInt(2, a);
+            preparedStatement.setInt(3, y);
             preparedStatement.executeUpdate();
         } catch (SQLException ex) {
             System.out.println(ex);
         }
     }
-    public boolean rech(int a) throws SQLException{
+    public boolean rech(int a, int x) throws SQLException{
            Statement stmt = connection.createStatement();
-         ResultSet RS = stmt.executeQuery("select * from orders where product_id like "+a);
+         ResultSet RS = stmt.executeQuery("select * from orders where user_id="+x+" product_id like "+a);
          if (RS.next())
          {//System.out.println("True");
              return true ;
@@ -98,5 +100,25 @@ public class OrderService {
          //System.out.println("False");
          return false ;
        }
-    
+      public ArrayList<Order> rechercherOrder( int x, Integer a) {
+         ArrayList<Order> part = new ArrayList<>();
+        String req = "select * from article where  user_id="+a+" and prduct_id like '%"+x+"%'";
+        System.out.println(req);
+        PreparedStatement preparedStatement;
+        try {
+            preparedStatement = connection.prepareStatement(req);
+           // preparedStatement.setInt(1, r);
+       
+            ResultSet resultSet = preparedStatement.executeQuery();
+            while (resultSet.next()) {
+               Order pa;
+              pa = new Order(resultSet.getInt("product_id"), resultSet.getInt("user_id"),resultSet.getInt("quantity"), resultSet.getInt("price"), resultSet.getInt("product_owner"));
+                System.out.println(pa);
+                part.add(pa);
+            }
+        } catch (SQLException ex) {
+            ex.printStackTrace();
+        }
+        return part;
+    }
 }
